Mudando tamanho de variáveis

+++ grafo.h
@@ -3,13 +3,13 @@
 
 class Vertice{
     public:
-        Vertice(int i);
+        Vertice(unsigned long i);
         
-        int getIndice();
-        int getGrau();
-        int getPai();
-        int getNivel();
-        int getMarcacao();
+        unsigned long getIndice();
+        unsigned long getGrau();
+        unsigned long getPai();
+        unsigned long getNivel();
+        uint8_t getMarcacao();
 
         void incrementaGrau();
         void setPai(int v);
@@ -18,18 +18,18 @@ class Vertice{
         void marca(int i=1);
     protected:
     private:
-        int indice;
-        int grau;
-        int marcacao;
-        int pai;
-        int nivel;
+        unsigned long indice;
+        unsigned long grau;
+        uint8_t marcacao;
+        unsigned long pai;
+        unsigned long nivel;
         
 };
 
 class Aresta{
     public:
-        Aresta(int v1, int v2);
-        Vertice* getVertice(int v);
+        Aresta(unsigned long v1, unsigned long v2);
+        Vertice* getVertice(unsigned long v);
     protected:
     private:
         Vertice *v[2];
@@ -37,41 +37,41 @@ class Aresta{
 
 class MatrizAdjacencias{
     public:
-        MatrizAdjacencias(int n);
+        MatrizAdjacencias(unsigned long n);
         MatrizAdjacencias(FILE *input);
 
-        void setAdjacencia(int v1, int v2);
+        void setAdjacencia(unsigned long v1, unsigned long v2);
 
-        void BFS(int origem);
-        void DFS(int origem);
+        void BFS(unsigned long origem);
+        void DFS(unsigned long origem);
 
         void componentesConexos();
 
     protected:
     private:
-        int n_vertices;
-        int n_arestas;
+        unsigned long n_vertices;
+        unsigned long n_arestas;
         Vertice **vertices;
-        int **adjacencias;
+        unsigned long **adjacencias;
 };
 
 class ListaAdjacencias{
     public:
-        ListaAdjacencias(int n);
+        ListaAdjacencias(unsigned long n);
         ListaAdjacencias(FILE *input);
 
-        void setAdjacencia(int v1, int v2);
+        void setAdjacencia(unsigned long v1, unsigned long v2);
 
-        void BFS(int origem);
-        void DFS(int origem);
+        void BFS(unsigned long origem);
+        void DFS(unsigned long origem);
 
         void componentesConexos();
     protected:
     private:
-        int n_vertices;
-        int n_arestas;
+        unsigned long n_vertices;
+        unsigned long n_arestas;
         Vertice **vertices;
-        list<int> **adjacencias;
+        list<unsigned long> **adjacencias;
 };
 
 #endif //GRAFO_H
